langchain_chat/requirements.txt
aiomysql==0.2.0
alembic==1.13.2
annotated-types==0.7.0
ansible==12.0.0
ansible-core==2.19.2
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
awsebcli==3.25.3
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.31.0
azure-mgmt-core==1.4.0
azure-mgmt-storage==21.2.1
azure-storage-blob==12.23.0
bcrypt==4.1.2
blessed==1.25.0
botocore==1.42.9
bs4==0.0.2
cement==2.10.14
certifi==2024.7.4
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==43.0.0
cycler==0.12.1
databases==0.9.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.3.1
distlib==0.3.9
dnspython==2.6.1
email_validator==2.1.2
executing==2.2.0
fabric==3.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.16.1
flake8==7.1.1
fonttools==4.57.0
fuzzywuzzy==0.18.0
git-filter-repo==2.47.0
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.6.15
idna==3.7
invoke==2.2.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.1
msrest==0.7.1
nest-asyncio==1.6.0
nodeenv==1.9.1
norminette==3.3.52
numpy==2.1.3
oauthlib==3.2.2
orjson==3.10.6
packaging==24.2
pandas==2.2.3
paramiko==4.0.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pi==0.1.2
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.3.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwdlib==0.2.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
PyNaCl==1.6.1
pynvim==0.6.0
pyparsing==3.2.3
pypi==2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.4.0
requests==2.32.3
requests-oauthlib==2.0.0
resolvelib==1.2.0
rich==13.7.1
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
termcolor==2.5.0
threadpoolctl==3.6.0
tld==0.13.1
tornado==6.4.2
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.30.3
uvloop==0.19.0
virtualenv==20.27.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wrapt==2.0.1
xsstrike==3.2.2
yamllint==1.37.1
